api/requirements.txt
# API dependencies
fastapi
uvicorn
python-multipart
pillow
torch==2.6.0
torchvision==0.21.0

# Force CPU version of PyTorch
-f https://download.pytorch.org/whl/cpu.html
